% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colors}
\alias{colors}
\title{colors}
\format{
An object of class \code{character} of length 5.
}
\usage{
colors
}
\description{
specify the fill color by job type fr plot functions
}
\keyword{datasets}
