% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCAM_JOB.R
\name{CON_JOB}
\alias{CON_JOB}
\title{CON_JOB}
\usage{
CON_JOB(GW_activity, method = NULL)
}
\arguments{
\item{GW_activity}{Output from GCAM_GW()}

\item{method}{Specify method for construction job.
"Net": in a given year, a technology pre-mature retirement and newly installed net out.
"Total": in a given year, a technology pre-mature retirement and newly installed can happen at the same time.}
}
\value{
A data frame of construction job by subsector
}
\description{
CON_JOB
}
\examples{
job.CON <- CON_JOB(GW_activity, "Net")
job.CON <- CON_JOB(GW_activity)
}
