% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCAM_EJ.R
\name{GCAM_EJ}
\alias{GCAM_EJ}
\title{GCAM_EJ}
\usage{
GCAM_EJ(prj)
}
\arguments{
\item{prj}{A list of queried outcome
This function takes +20 seconds for a single scenario}
}
\value{
A data frame with state-level power output from capacity activity
}
\description{
GCAM_EJ
}
\examples{
EJ_activity <- GCAM_EJ(prj)
}
