% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCAM_JOB.R
\name{GCAM_JOB}
\alias{GCAM_JOB}
\title{GCAM_JOB}
\usage{
GCAM_JOB(GW_activity, method)
}
\arguments{
\item{GW_activity}{Output from GCAM_GW()}

\item{method}{Specify method for construction job.
"Net": in a given year, a technology pre-mature retirement and newly installed net out.
"Total": in a given year, a technology pre-mature retirement and newly installed can happen at the same time.}
}
\value{
A data frame of all power sector direct job by subsector
}
\description{
GCAM_JOB
}
\examples{
JOB_activity <- GCAM_JOB(GW_activity, "Net")
JOB_activity <- GCAM_JOB(GW_activity, "Total")
}
