% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hydro_cf}
\alias{hydro_cf}
\title{hydro_cf}
\format{
A data frame with \code{49} rows and \code{3} variables:
}
\source{
Hall et al., 2003. Table 8, exported as GODEEP/Assumptions/hydro_CF_USA.csv
state level monthly capacity factor
}
\usage{
hydro_cf
}
\description{
This data provides a state-level annual capacity factor of hydro power
}
\examples{
data(hydro_cf)
}
\keyword{datasets}
