% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{MAP_JOB}
\alias{MAP_JOB}
\title{MAP_JOB}
\usage{
MAP_JOB(JOB_activity, year)
}
\arguments{
\item{JOB_activity}{Output of GCAM_JOB()}

\item{year}{Year of interest, a numeric}
}
\value{
A map of state level total direct power sector job of a given year
}
\description{
MAP_JOB
}
\examples{
MAP_JOB(JOB_activity, 2050)
}
