% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapUS52Compact}
\alias{mapUS52Compact}
\title{mapUS52Compact}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 51 rows and 7 columns.
}
\source{
https://github.com/JGCRI/rmap/tree/main/data
}
\usage{
mapUS52Compact
}
\description{
This data presents the spatial information of USA states
}
\examples{
data(mapUS52Compact[, 1:6])
}
\keyword{datasets}
