% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OM_cost}
\alias{OM_cost}
\title{OM_cost}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8074 rows and 7 columns.
}
\source{
GCAM data
}
\usage{
OM_cost
}
\description{
This data technology level OM fixed and variable cost information
}
\examples{
data(OM_cost)
}
\keyword{datasets}
