% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCAM_JOB.R
\name{OM_JOB}
\alias{OM_JOB}
\title{OM_JOB}
\usage{
OM_JOB(GW_activity)
}
\arguments{
\item{GW_activity}{Output from GCAM_GW()}
}
\value{
A data frame of OM job by subsector
}
\description{
OM_JOB
}
\examples{
job.OM <- OM_JOB(GW_activity)
}
