% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PLOT_JOB}
\alias{PLOT_JOB}
\title{PLOT_JOB}
\usage{
PLOT_JOB(JOB_activity)
}
\arguments{
\item{JOB_activity}{Output from GCAM_JOB()}
}
\value{
A trend plot of job by fuel technology and job type at the national level
}
\description{
PLOT_JOB
}
\examples{
PLOT_JOB(JOB_activity)
}
