% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PLOT_JOB_TYPE}
\alias{PLOT_JOB_TYPE}
\title{PLOT_JOB_TYPE}
\usage{
PLOT_JOB_TYPE(JOB_activity)
}
\arguments{
\item{JOB_activity}{Output of GCAM_JOB()}
}
\value{
A trend plot of job by job type at the national level
}
\description{
PLOT_JOB_TYPE
}
\examples{
PLOT_JOB_TYPE(JOB_activity)
}
