% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SCurve}
\alias{SCurve}
\title{SCurve}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 395127 rows and 7 columns.
}
\source{
GCAM assumptions used in GCAM-USA 7.1; see scripts for details
}
\usage{
SCurve
}
\description{
This data provides the shut-down assumptions by fuel and technology at the state level
}
\details{
Follow package plutus to gather the assumptions
}
\examples{
data(SCurve)
}
\keyword{datasets}
