% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theme_leg}
\alias{theme_leg}
\title{theme_leg}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 8.
}
\usage{
theme_leg
}
\description{
specify plot theme_leg
}
\keyword{datasets}
